+++ src/data/bucket/symbol_bucket.ts
@@ -348,6 +348,7 @@ export class SymbolBucket implements Bucket {
     canOverlap: boolean;
     sortedAngle: number;
     featureSortOrder: number[];
+    maxHeightOffset: number;
 
     collisionCircleArray: number[];
 
@@ -374,6 +375,7 @@ export class SymbolBucket implements Bucket {
         this.sourceLayerIndex = options.sourceLayerIndex;
         this.hasDependencies = false;
         this.hasRTLText = false;
+        this.maxHeightOffset = 0;
         this.sortKeyRanges = [];
 
         this.collisionCircleArray = [];